// src/interfaces/IPatient.ts
export default interface IPatient {
	birthday?: string;
	joined_at?: string;
	mother_name?: string;
	substances?: string;
	criminal_record?: string;
	cep?: string;
	city?: string;
	cpf?: string;
	created_at?: string;
	gender?: string;
	id?: string;
	managers_id?: string;
	managers_name?: string;
	name?: string;
	number_house?: string;
	patient_place?: string;
	profile_photo?: string;
	rg?: string;
	status?: string;
	street?: string;
	uf?: string;
	updated_at?: string;
	description?: string;
	progress_type?: string;
}